Add API to connect to Imagga and /images endpoint
Full implementation for ImaggaConnector added that supports submission
of an image URL which will be sent for recognition.

Added ImagesController with endpoint /images that will execute a
POST request.

Added class Constants in new module util which will be used to hold
constants relevant to the implementation of the project.

Modify build.yml