Major additions
Added some dependencies to pom.xml

Modify connect() method of Connector and ImaggaConnector to have return
type String and to return the API JSON response as a String.

Add ImageDAO which will be responsible for CRUD operations
for the images.

Add PostgresDAO which opens and closes DB connection.

Add ImageData DTO which holds info for the image which
will be submitted to the DB via ImageDAO and PostgresDAO.

Add ImageModel (implementation TBI)

Add util ChecksumCalculator to generate a checksum for the URL
to be stored in the DB table images. It is for optimized search
purposes.

Minor modifications to Constants class to hold DB connection params.

Add util ImageDimensions which returns image width and height based
on the URL passed.

Add util ResponseParser to take the String JSON reposnse from the 3rd party
API, convert them to JSONObject and strip the response from
unecessary data.