Finished basic persistence
Finish ImageDAO so that it can persist Image data and tags to the database
Additionally, it also persists the relation between an image and a tag
together with the confidence level in the tag.

Modify ImageController to call ImageModel.

Add implementation to ImageModel to open connection to DB,
persist data, and close the conncetion.