Finish "/" endpoint
Finish "/" endpoint logic where user submits image url and the
program checks if image already exists. If it does it fetches the data
and stores it in a ImageData object. If not it sends it to Imagga API
for analysis and stores the reposnse in the DB and then fetches it
to store it in an ImageData object.

Additionally, parameter noCache implemented to indicate whether we
wish to directly overwrite existing images.